Food Classification/requirements.txt
anthropic==0.42.0
google-cloud-storage==2.18.2
numpy==2.1.3
pandas==2.2.3
openai==1.57.4
google-generativeai==0.8.3